Moved some code into roi extractor.
